integrate kmp big decimals in fund service.

+++ service/fund/fund-service/src/main/kotlin/ro/jf/funds/fund/service/mapper/TransactionMapper.kt
@@ -52,8 +52,7 @@ fun TransactionRecord.toTO() = TransactionRecordTO(
     id = id,
     accountId = accountId,
     fundId = fundId,
-    // TODO(Johann) not ok
-    amount = amount.toDouble(),
+    amount = amount,
     unit = unit,
     labels = labels,
 )
